+++ scripts/macros/features/classFeatures.js
@@ -1,9 +1,11 @@
 ﻿import { barbarian } from './classFeatures/barbarian.js';
 import { rogue } from './classFeatures/rogue.js';
 import { cleric } from './classFeatures/cleric.js';
+import { bard } from './classFeatures/bard.js';
 
 export let classFeatures = {
 	'barbarian': barbarian,
 	'rogue': rogue,
-	'cleric': cleric
+	'cleric': cleric,
+	'bard': bard
 }