+++ scripts/macros/spells.js
@@ -9,6 +9,9 @@ import { spiritualWeapon } from './spells/spiritualWeapon.js';
 import { spiritGuardians } from './spells/spiritGuardians.js';
 import { auraOfVitality } from './spells/auraOfVitality.js';
 import { huntersMark } from './spells/huntersMark.js';
+import { hungerHadar } from './spells/hungerHadar.js';
+import { conjureCloud } from './spells/cloudOfDaggers.js';
+import { heatMetal } from './spells/heatMetal.js';
 
 export let spells = {
 	'hex': hex,
@@ -21,5 +24,8 @@ export let spells = {
 	'spiritualWeapon': spiritualWeapon,
 	'spiritGuardians': spiritGuardians,
 	'auraOfVitality': auraOfVitality,
-	'huntersMark' : huntersMark
+	'huntersMark': huntersMark,
+	'hungerHadar': hungerHadar,
+	'cloudOfDaggers': conjureCloud,
+	'heatMetal': heatMetal
 }