+++ scripts/macros/spells/spiritualWeapon.js
@@ -3,38 +3,44 @@
 export async function spiritualWeapon(args, summonImage, summonTokenImage) {
 	let lastArg = args[args.length-1];
 	let token = canvas.tokens.get(lastArg.tokenId);
-	let summonData = await ggHelpers.getItemFromCompendium('garhis-grotto.gg-actor-blueprints', 'Spiritual Weapon', false);
-	if (!summonData) return;
 	let summonName = `Spiritual Weapon - ${lastArg.actor.name}`;
-	// TODO: Prompt the user to select art maybe
-	summonData.name = summonName;
-	summonData.img = summonImage;
-	summonData.prototypeToken.name = summonName;
-	summonData.prototypeToken.texture.src = summonTokenImage;
-	let summon = await Actor.create(summonData);
-
-	let summonId = await warpgate.spawn(summonName, {}, {}, {'controllingActor': lastArg.actor});
-	console.log(summonId);
+	let spellMod = lastArg.actor.system.abilities[lastArg.actor.system.attributes.spellcasting].mod;
+	let attackBonus = spellMod + lastArg.actor.system.attributes.prof + lastArg.actor.system.bonuses.msak.attack;
+	let damageBonus = spellMod + lastArg.actor.system.bonuses.msak.damage;
+	let damage = { parts: [[`${Math.floor(lastArg.castData.castLevel/2)}d8+${damageBonus}`, "force"]], versatile: "" };
+	let attackData = await ggHelpers.getItemFromCompendium('garhis-grotto.gg-item-blueprints', 'Spiritual Weapon Attack', false);
+	attackData.system.damage = damage;
+	attackData.system.attackBonus = attackBonus;
+	attackData.img = summonImage;
+	let summonUpdates = {
+		'token': {
+			'name': summonName,
+			'texture.src': summonTokenImage
+		},
+		'actor': {
+			'name': summonName,
+			'img': summonImage,
+		},
+		'embedded': {
+			'Item': {
+				[attackData.name]: attackData
+			},
+		}
+	}
+	let summonId = await warpgate.spawn('GG - Spiritual Weapon', summonUpdates, {}, {'controllingActor': lastArg.actor});
 	if (!summonId) {
-		await Actor.deleteDocuments([summon.id]);
 		ui.notifications.warn('Warpgate Summoning Aborted');
 		return;
 	}
 	const summonedToken = game.canvas.tokens.get(summonId[0]);
 	const summonTokenUuid = summonedToken.document.uuid;
-	let damage = { parts: [[`${Math.floor(lastArg.castData.castLevel/2)}d8+@mod`, "force"]], versatile: "" };
-	let attackData = await ggHelpers.getItemFromCompendium('garhis-grotto.gg-item-blueprints', 'Spiritual Weapon Attack', false);
-	attackData.system.damage = damage;
-	attackData.img = summonImage;
 	async function effectMacro() {
 		await warpgate.revert(token.document, 'Spiritual Weapon');
-		await Actor.deleteDocuments([origin.id]);
 	}
 	let sourceEffectData = {
 		'label': 'Spiritual Weapon',
 		'icon': summonImage,
 		'changes': [ { 'key': 'flags.dae.deleteUuid', 'mode': CONST.ACTIVE_EFFECT_MODES.ADD, 'value': [summonTokenUuid] } ],
-		'origin': summon.uuid,
 		'duration': { 'seconds': 60, rounds:10 },
 		'flags': {
 			'dae': {
@@ -47,6 +53,9 @@ export async function spiritualWeapon(args, summonImage, summonTokenImage) {
 			}
 		}
 	}
+	attackData.system.proficient = true;
+	attackData.system.attackBonus = "";
+	attackData.system.damage = { parts: [[`${Math.floor(lastArg.castData.castLevel/2)}d8+@mod`, "force"]], versatile: "" };
 	let actorUpdates = {
 		'embedded': {
 			'Item': {