+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+# 0.0.12
+* Added Template Macro as a hard requirement
+* Removed Advanced Macros requirement
+* Hunger of Hadar
+* Reworked Spiritual Weapon summoning
+* Cloud of Daggers
+* Heat Metal
+* Unsettling Words
+* Auto Cast Reckless Attack
+
 # 0.0.11
 Added Path Features and Writeup for Nocturne
 Bugfix for Hunter's Mark not removing effect when transferred